requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aioshutil==1.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
litellm==1.53.2
markdownify==0.13.1
marker-pdf==1.0.1
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
onnxruntime==1.20.1
openai==1.56.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdftext==0.3.19
pillow==10.4.0
propcache==0.2.1
protobuf==5.29.0
py-zerox==0.0.7
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
surya-ocr==0.7.0
sympy==1.13.1
tabled-pdf==0.1.7
tabulate==0.9.0
texify==0.2.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.3
zipp==3.21.0
